requirements.txt
faiss-cpu==1.8.0.post1
numpy>=1.26.4,<2.0
scikit-learn>=1.3.2,<2.0
tqdm>=4.66.4
ujson>=5.10.0
pyyaml>=6.0.1
sentence-transformers>=2.6.1
